Add files via upload
Classes is a pretty clean dataset consisting of near all the classes which show up in TWI, hypothetical or otherwise.

Skills is a clean dataset consisting of near all the skills that occur in TWI. 

I generated these with TWI Skills and Classes Scraper-Copy1.ipynb and I realized I hate regexes. 

They should be up to date circa Interlude: Death and Stitches. I'm not sure if they contain the secret chapters though. I can check this by looking through the chapter_links variable and just checking if the secret chapters are in there.